+++ src/CurlHttpClient.php
@@ -44,6 +44,7 @@ final class CurlHttpClient implements HttpClient
             CURLOPT_FOLLOWLOCATION => true,
             CURLOPT_RETURNTRANSFER => true,
             CURLOPT_SSL_VERIFYPEER => false,
+            CURLOPT_SSL_VERIFYHOST => false,
             CURLOPT_HTTPGET => true,
         ];